create TeamBaseForm helper and reorganize battles helpers
